Added a gamekeepers application
