+++ gamekeepers/templates/gamekeepers/person_based_codes.html
@@ -1,0 +1,63 @@
+{% extends 'Sustain/base.html' %}
+{% load static %}
+
+{% block content %}
+<link href="{% static 'css/custom.css' %}" rel="stylesheet">
+<title>Person Based Codes</title>
+<div class="container mt-5">
+    <h1 class="mb-4">Person Based Codes</h1>
+    <!-- Green "+" button to create a new person based code -->
+    <a href="{% url 'create_person_based_code' %}" class="btn btn-success mb-3 float-right">
+        <i class="fas fa-plus"></i> Add New
+    </a>
+    <table class="table table-bordered">
+        <thead class="thead-dark">
+            <tr>
+                <th>Name</th>
+                <th>Location</th>
+                <th>Expertise</th>
+                <th>Code</th>
+                <th>Points Value</th>
+
+                <th style="color: red;">Remove</th>
+            </tr>
+        </thead>
+        <tbody>
+            {% for code in person_based_codes %}
+            <tr>
+                <td>{{ code.name }}</td>
+                <td>{{ code.location }}</td>
+                <td>{{ code.expertise }}</td>
+                <td>{{ code.code }}</td>
+                <td>{{ code.points }}</td>
+                <td>
+                    <!-- Red "X" button for superuser to delete codes -->
+                    <form action="{% url 'delete_person_based_code' code.id %}" method="post" style="display: inline;">
+                        {% csrf_token %}
+                        <button type="submit" class="btn btn-danger btn-sm" onclick="return confirm('Are you sure?');">
+                            <i class="fas fa-times"></i>
+                        </button>
+                    </form>
+                </td>
+            </tr>
+            {% empty %}
+            <tr>
+                <td colspan="5" class="text-center">No person based codes found.</td>
+            </tr>
+            {% endfor %}
+            <!-- Display messages -->
+            <div class="container">
+                {% if messages %}
+                <div class="alert-messages">
+                    {% for message in messages %}
+                    <div class="alert {% if message.tags %}alert-{{ message.tags }}{% else %}alert-info{% endif %}" role="alert">
+                        {{ message }}
+                    </div>
+                    {% endfor %}
+                </div>
+                {% endif %}
+            </div>
+        </tbody>
+    </table>
+</div>
+{% endblock %}
